Better analysis feed

+++ inc/oanda/request_builder.h
@@ -7,16 +7,16 @@
 
 #include "analysis/enumations.h"
 
-enum RISKI_ERROR_CODE oanda_v20_v3_accounts(char* host, char* api_key,
-                                            char** res);
-enum RISKI_ERROR_CODE oanda_v20_v3_accounts_instruments(char* host,
-                                                        char* api_key,
-                                                        char** res,
-                                                        char* account_id);
+enum RISKI_ERROR_CODE oanda_v20_v3_accounts(char *host, char *api_key,
+                                            char **res);
+enum RISKI_ERROR_CODE oanda_v20_v3_accounts_instruments(char *host,
+                                                        char *api_key,
+                                                        char **res,
+                                                        char *account_id);
 
-enum RISKI_ERROR_CODE oanda_v20_v3_accounts_pricing(char* host, char* api_key,
-                                                    char** instrument_name,
-                                                    size_t num_instruments,
-                                                    char* account_id,
-                                                    char** res);
+enum RISKI_ERROR_CODE oanda_v20_v3_accounts_pricing(char *host, char *api_key,
+                                                    char **instrument_name,
+                                                    int num_instruments,
+                                                    char *account_id,
+                                                    char **res);
 #endif
